Reduce up-front imports, dedup introspection code, ping returns method list
This improves baseline startup speed for the client.

The ping change in particular makes it possible for the service to load or
generate code after startup, which the client isn't aware of until the .ping()
during .connect(). This further reduces client startup times and allows for
modularity down the line.

Knwon buglet: the CLI `help` command currently does not know about lazy-loaded
API functions.